refactor(agent): extract provider/URL logic from Factory into ProviderSpec
