+++ AGENTS.md
@@ -66,7 +66,7 @@ This file is your **map**. Read the relevant doc before starting any task.
 ### Agent Service Rules (apply when touching `lib/zaq/agent/`)
 
 - **Use existing infrastructure before building new paths** — before implementing any LLM call, agent lifecycle, or response builder, confirm whether `Factory`, `Executor`, `Outgoing`, or `History` already covers the case. If it does, use or extend it. Never create a parallel path. See `docs/services/agent.md` for the Entry Point Decision Tree.
-- **Provider/URL logic has one home** — all AI provider credential resolution and endpoint URL formatting belongs exclusively in `get_ai_provider_credential/1` and `Zaq.Agent.Factory`. All other modules (`ServerManager`, `Pipeline`, `Answering`, etc.) receive pre-built model spec objects and must never construct or inspect provider URLs or credentials directly.
+- **Provider/URL logic has one home** — all provider-atom normalisation, fixed-URL detection, and base-URL injection live in `Zaq.Agent.ProviderSpec`. Credential resolution lives in `get_ai_provider_credential/1`. Model spec assembly (calling both) lives in `Zaq.Agent.Factory`. All other modules (`ServerManager`, `Pipeline`, `Answering`, etc.) receive pre-built model spec objects and must never construct or inspect provider URLs or credentials directly.
 - **The default answering agent is not a special case** — it is a configured agent with default values. `ServerManager` must not branch on agent type. Any logic that applies only to "the answering agent" is a smell: either it belongs to all agents (move it to the general path) or it should not exist.
 - **`nil` person_id is never an implicit permission grant** — if a function receives `person_id: nil` with no explicit `skip_permissions: true` in context, it must return only public data. Explicit admin access must be opt-in. Never derive permissions from nil. See `docs/services/agent.md` security note.
 - **`ServerManager` state must be minimal** — if a state variable exists only to trigger future behavior, use `Process.send_after/3` instead. Ask: "would this variable still be needed if I used a timer?" If no, use the timer.
